Chore: (`demo-app`) - Refactor and enhance accordion and channels interactions
Refactoring of accordion and channels interactions was done to improve code readability as well as functionality. The accordion interactions are now more intuitive with functions to handle events such as keypress, click, focus and blur, thereby improving the user interaction. Also, command hooks for editing the title, adding and removing items have been implemented to provide a more interactive UX in the channels interactions. This change will improve the overall user experience.