Tweak: (`commander`) - Update callback now passing currentState
The callback function in setState across various files was updated to return the current state. Previously, the setState callback function did not take any arguments, however, for improving the usability and add the functionality to handle the current state within the callback, the callback is updated to take `state` as argument. Now, callback can utilize the current `state` for any further operations or checks.