Tweak: (`commander`) - Add global event emitter to CommandBase class
EventEmitter was moved from import as a type to import as a class to allow the creation of a new EventEmitter instance. This was created as a 'global' within the CommandBase abstract class. The aim is to provide additional functionality for CommandBase classes, allowing them to hook into global events and emit events on a global scope. Two new methods were added, 'hook' and 'global' to respectively enable callback registration and provide access to global settings and instances within an instance of a CommandBase subclass.