Fix: (`api/commander`) - Refactor lifecycle handling, move handlers from context to instance
Previous implementation relied on storing lifecycle handlers in the context. This has been refactored to store lifecycle handlers directly on the instance of a component wrapped with the `with-commands` HOC. The lifecycle handlers are now under the `$$commander.lifecycleHandlers` property instead of `context.internalHandlers`. This adjustment improves handling of lifecycle events by avoiding potential issues with context propagation and improving encapsulation of the handlers inside the component instances.