New: (`api`) - Package
Refactored the codebase by moving the api-related files (api-core, index, and api-component) from the 'demo-app' directory to the new 'api' package. This provides a better organization of the code, promoting reusability and scalability. The api now stands as a separate module that can be used across multiple apps.

In addition, 'api' as a development dependency was added to the package.json of demo-app. This indicates that the demo-app is dependent on the 'api' package during the development stage. Code within api-channels-module was also edited to reflect these changes.

New files like package.json, tsconfig.json, and tsconfig.eslint.json were created within packages/api directory to configure the api package appropriately.

Additionally, a few unused variables were cleaned up, and the code was enhanced with more detailed commenting for better readability.