Fix: (`demo-app`) - APIChannelsModule fix and enhance autosave feature
Added globalUnhook() to the CommandBase class in command-base.ts for unhooking event listeners. This allows us to unhook listeners when they are no longer needed to prevent memory leaks and unneeded executions.

Also updated autosave functionality in api-channels-module.ts. Now channels' data auto-saves with debounce, which only activates when necessary and reduces the number of unnecessary save requests to the server. Furthermore, added cleanup for the event listeners and autosave handlers when the component is unmounted. These changes enhance the efficiency of autosaving and clean up actions once they're done.