New: (`demo-app`) - Add smooth scrolling feature to channel table
New SVG symbol 'ArrowSkinnyRight' was added to the UI symbols. It was utilized in a smooth scrolling feature in the channel-item-table component. Smooth scrolling is now possible to left and right by clicking the arrow. It also adjusts its orientation based on the scroll direction. UI updates were also made to accommodate the new feature. Additionaly, channels list table was updated to render only channels that have some budget allocation and alert was added if none of them have one.